Skip tests requiring iptables.
Tests running `sudo` are problematic to run locally during development,
as you can't just run the tests - you need to watch them and wait for
sudo prompt. In this test, it would be better to use some kind of proxy.

As the test was already marked xfail, skip it to make development easier.